* 필터 충돌 해결  * 반환값 오류 및 이중 소멸자 오류 수정

+++ Sparta_Cpp_Assginment/Character.cpp
@@ -128,7 +128,7 @@ int Character::getAttack() const
 	return attack;
 }
 
-int Character::getInventorySize() const
+size_t Character::getInventorySize() const
 {
 	return inventory.size();
 }
